+++ staging/src/k8s.io/apiextensions-apiserver/pkg/helpers/v1beta1/helpers.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package helpers
+package v1beta1
 
 import (
 	"fmt"
@@ -103,8 +103,8 @@ func IsCRDConditionEquivalent(lhs, rhs *apiextensions.CustomResourceDefinitionCo
 	return lhs.Message == rhs.Message && lhs.Reason == rhs.Reason && lhs.Status == rhs.Status && lhs.Type == rhs.Type
 }
 
-// CRDHasFinalizer returns true if the finalizer is in the list
-func CRDHasFinalizer(crd *apiextensions.CustomResourceDefinition, needle string) bool {
+// HasFinalizer returns true if the finalizer is in the list
+func HasFinalizer(crd *apiextensions.CustomResourceDefinition, needle string) bool {
 	for _, finalizer := range crd.Finalizers {
 		if finalizer == needle {
 			return true
@@ -114,8 +114,8 @@ func CRDHasFinalizer(crd *apiextensions.CustomResourceDefinition, needle string)
 	return false
 }
 
-// CRDRemoveFinalizer removes the finalizer if present
-func CRDRemoveFinalizer(crd *apiextensions.CustomResourceDefinition, needle string) {
+// RemoveFinalizer removes the finalizer if present
+func RemoveFinalizer(crd *apiextensions.CustomResourceDefinition, needle string) {
 	newFinalizers := []string{}
 	for _, finalizer := range crd.Finalizers {
 		if finalizer != needle {
